Fix bug related to question 704965
Bug made the program to try loop-induced process even when other errors,
not related to amplitude generation occurred, so that it fails by saying
that the model is not NLO compatible even when this does not make sense.
Fixed by setting a model-related flag regarding if the model allows
loop-induced computations and by using the InvalidCmd and
NoDiagramException error properly.