Add computation of J-factors and profiles in a separate file
J-factors computation has a new argument 'strategy' to be set
according to the kind of mask one wants to compute.
Moreover, it is possible specify the r_max parameter as the
maximum value of the line-of-sight to integrate over.
The line experiments list has been moved as a global variable
so that it is not re-initialized at every new run of the program.
The regions of interests are specified through a new class, to
keep into account different parameters and to potentially
have a different mask for each ROI, according to the experiment.
The maddm_card has been updated with the strategy and r_max
parameter for the template experiment.