Fix fast mode not considering specified final states. Pass to 3.2.11
When the user specifies only certain final states for indirect detection
through the command 'generate' or 'add', fast mode should pick only them
and do the various computations only basing on those. This was not the
case as the maddm.out file was simply parsed and everything was stored
(because it uses the same final states as relic density uses).
The behaviour was fixed by storing which final states were explicitly
asked by the user, and using them to save only those sigmav.