Add support for process names handling
When processes are generated, their human-readable name is matched against
a parsing-friendly version. Now there is full support for all different
processes, including various and several decay chains.
Moreover, a new class StrProcess allows one to handle the processes names, finding
the final states (after the development of the decay chains) and transforming
the parsing-friendly string in a better-looking format if needed.
Finally, I went back to the original process labels in terminal and file outputs
because of the fact there are still processes that would appear with vl, q, l, ecc
so that they can not be converted to pdg codes. So now it is consistent with them.
The pdg codes process syntax utilizes a dot '.' to separate different particles,
a '>' to indicate a process and encloses all decay chains in parenthesis, separating
them with commas ','. I removed the previuos syntax that used '(<pdg>).(<pdg>)'
to separate pdg codes, and now everything is handled through the class StrProcess.